+++ types/gauge-chart/index.d.ts
@@ -1,0 +1,1 @@
+export { RowanGaugeChart } from "./gauge-chart.js";